requirements.txt
aiosqlite==0.20.0
datasketch==1.6.4
sqlvalidator==0.0.20
func-timeout==4.3.5
grpcio-status==1.62.2
httptools==0.6.1
ipykernel==6.29.4
langchain-anthropic==0.1.16
langchain-chroma==0.1.2
langchain-community==0.2.6
langchain-google-genai==1.0.9
langchain_google_vertexai==1.0.8
google-cloud-aiplatform==1.63.0
langchain-openai==0.1.12
langgraph==0.1.4
pandas==2.2.1
pip-chill==1.0.3
pipreqs==0.5.0
python-dotenv==1.0.1
sentence-transformers==2.7.0
sqlglot==23.3.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
filelock==3.15.4
faiss-cpu==1.8.0
datasets==2.21.0
pyyaml==6.0.2
# MySQL Support
pymysql==1.1.0
dbutils==3.1.0
cryptography==42.0.5  # Required for secure MySQL connections